split sitemap generation
