+++ .github/workflows/fetch-external-hreflangs.yml
@@ -1,9 +1,12 @@
-name: Generate sitemap
+name: Fetch external hreflangs
 
 on:
-  schedule:
-    - cron: '0 0,12 * * *'  # This cron expression runs at 00:00 and 12:00 UTC every day
+  # schedule:
+  #   - cron: '0 0,12 * * *'  # This cron expression runs at 00:00 and 12:00 UTC every day
   workflow_dispatch:
+
+permissions:
+  contents: write
 
 jobs:
   run-script:
@@ -22,8 +25,8 @@ jobs:
       run: npm install
       working-directory: ./tools/sitemap-generation
 
-    - name: Generate sitemaps
-      run: node index.js
+    - name: Fetch external hreflangs
+      run: node fetch-external-hreflangs.js
       working-directory: ./tools/sitemap-generation
 
     - name: Configure Git
@@ -31,10 +34,21 @@ jobs:
         git config --global user.name 'sitemap-generation[bot]'
         git config --global user.email 'sitemap-generation[bot]@users.noreply.github.com'
 
-    - name: Commit and push changes
+    - name: Check for changes
+      id: changes
       run: |
         git add .
-        git commit -m "Update sitemaps"
+        if git diff-index --quiet HEAD --; then
+        echo "No changes to commit"
+        echo "changes=false" >> $GITHUB_OUTPUT
+        else
+        echo "changes=true" >> $GITHUB_OUTPUT
+        fi
+
+    - name: Commit and push changes
+      if: steps.changes.outputs.changes == 'true'
+      run: |
+        git commit -m "Update Sitemaps"
         git push
       env:
         GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
