+++ tools/sitemap-generation/package.json
@@ -5,7 +5,8 @@
   "main": "index.js",
   "type": "module",
   "scripts": {
-    "generate": "node index.js"
+    "generate": "node generate-sitemap.js",
+    "fetch-external": "node fetch-external-hreflangs.js"
   },
   "dependencies": {
     "axios": "^1.8.1",
